Rename joypad_switch.rs to joypad_nintendo.rs.
