+++ bindings/rust/src/joypad_nintendo.rs
